tests: move `State` implementation into state.rs
