+++ tests/steps/address.rs
@@ -1,6 +1,7 @@
+use crate::error;
 use crate::error::Error;
+use crate::state::{State, Step, StepStatus};
 use crate::steps::admin::IndexCosmogony;
-use crate::{error, State, Step, StepStatus};
 use async_trait::async_trait;
 use common::document::ContainerDocument;
 use config::Config;
